requirements.txt
aiobotocore==2.15.2
aiofiles==23.2.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.9.2
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-fab==1.1.1
apache-airflow-providers-ftp==3.9.1
apache-airflow-providers-google==10.26.0
apache-airflow-providers-http==4.11.1
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
bleach==6.2.0
blinker==1.8.2
botocore==1.35.16
cachelib==0.9.0
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
configparser==7.0.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.8
db-dtypes==1.3.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.9
dnspython==2.6.1
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
eval_type_backport==0.2.0
filelock==3.16.1
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.10.0
gcloud-aio-auth==5.3.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.3.0
gcsfs==2024.10.0
google-ads==25.1.0
google-analytics-admin==0.23.2
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.73.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-automl==2.14.1
google-cloud-batch==0.17.31
google-cloud-bigquery==3.27.0
google-cloud-bigquery-datatransfer==3.17.1
google-cloud-bigtable==2.27.0
google-cloud-build==3.27.1
google-cloud-compute==1.22.0
google-cloud-container==2.54.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.23.0
google-cloud-dataflow-client==0.8.14
google-cloud-dataform==0.5.13
google-cloud-dataplex==2.4.0
google-cloud-dataproc==5.15.1
google-cloud-dataproc-metastore==1.16.1
google-cloud-dlp==3.25.1
google-cloud-kms==3.1.1
google-cloud-language==2.15.1
google-cloud-logging==3.11.3
google-cloud-memcache==1.10.1
google-cloud-monitoring==2.23.1
google-cloud-orchestration-airflow==1.15.1
google-cloud-os-login==2.15.1
google-cloud-pubsub==2.27.1
google-cloud-redis==2.16.1
google-cloud-resource-manager==1.13.1
google-cloud-run==0.10.13
google-cloud-secret-manager==2.21.1
google-cloud-spanner==3.50.1
google-cloud-speech==2.28.1
google-cloud-storage==2.18.2
google-cloud-storage-transfer==1.13.1
google-cloud-tasks==2.17.1
google-cloud-texttospeech==2.21.1
google-cloud-translate==3.18.0
google-cloud-videointelligence==2.14.1
google-cloud-vision==3.8.1
google-cloud-workflows==1.15.1
google-crc32c==1.6.0
google-re2==1.1.20240601
google-resumable-media==2.7.2
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.68.0
grpcio-gcp==0.2.2
grpcio-status==1.68.0
gspread==6.1.4
gunicorn==22.0.0
h11==0.14.0
hatch==1.13.0
hatchling==1.26.3
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
immutabledict==4.2.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jmespath==1.0.1
json-merge-patch==0.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kaggle==1.6.17
keyring==25.5.0
lazy-object-proxy==1.10.0
limits==3.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==24.20.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.3.0
multidict==6.0.5
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
ordered-set==4.1.0
packaging==24.2
pandas==2.1.4
pandas-gbq==0.24.0
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.5.0
praw==7.7.1
prawcore==2.4.0
prison==0.2.1
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydata-google-auth==1.9.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.18.1
rsa==4.9
s3fs==2024.10.0
setproctitle==1.3.3
setuptools==70.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
sqlalchemy-bigquery==1.12.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.7.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.2
tabulate==0.9.0
tenacity==8.4.1
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
tomli_w==1.1.0
tomlkit==0.13.2
tqdm==4.67.0
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.0
userpath==1.9.2
uv==0.5.4
virtualenv==20.27.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
zstandard==0.23.0
